+++ app/src/routes/api/v1/area.router.js
@@ -17,6 +17,14 @@ const serializeObjToQuery = (obj) => Object.keys(obj).reduce((a, k) => {
     return a;
 }, []).join('&');
 
+const getHostForPaginationLink = (ctx) => {
+    if ('referer' in ctx.request.header) {
+        const url = new URL(ctx.request.header.referer);
+        return url.host;
+    }
+    return ctx.request.host;
+};
+
 class AreaRouter {
 
     static async getAll(ctx) {
@@ -32,7 +40,7 @@ class AreaRouter {
         delete clonedQuery.ids;
         const serializedQuery = serializeObjToQuery(clonedQuery) ? `?${serializeObjToQuery(clonedQuery)}&` : '?';
         const apiVersion = ctx.mountPath.split('/')[ctx.mountPath.split('/').length - 1];
-        const link = `${ctx.request.protocol}://${ctx.request.host}/${apiVersion}${ctx.request.path}${serializedQuery}`;
+        const link = `${ctx.request.protocol}://${getHostForPaginationLink(ctx)}/${apiVersion}${ctx.request.path}${serializedQuery}`;
 
         ctx.body = AreaSerializer.serialize(areas, link);
     }