+++ app/test/e2e/v1/get-areas-pagination.spec.js
@@ -1,5 +1,6 @@
 const nock = require('nock');
 const chai = require('chai');
+const config = require('config');
 const Area = require('models/area.model');
 
 const { createArea, mockGetUserFromToken } = require('../utils/helpers');
@@ -10,7 +11,7 @@ chai.should();
 
 const requester = getTestServer();
 
-describe('V1 - Get areas tests', () => {
+describe('V1 - Get areas', () => {
 
     before(async () => {
         if (process.env.NODE_ENV !== 'test') {
@@ -20,6 +21,41 @@ describe('V1 - Get areas tests', () => {
 
     beforeEach(async () => {
         await Area.deleteMany({}).exec();
+    });
+
+    describe('Test pagination links', () => {
+        it('Get areas without referer header should be successful and use the request host', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v1/area`)
+                .set('Authorization', 'Bearer abcd');
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/area?page[number]=1&page[size]=1000`);
+            response.body.links.should.have.property('prev').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/area?page[number]=1&page[size]=1000`);
+            response.body.links.should.have.property('next').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/area?page[number]=1&page[size]=1000`);
+            response.body.links.should.have.property('first').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/area?page[number]=1&page[size]=1000`);
+            response.body.links.should.have.property('last').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/area?page[number]=1&page[size]=1000`);
+        });
+
+        it('Get areas with referer header should be successful and use that header on the links on the response', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v1/area`)
+                .set('referer', `https://potato.com/get-me-all-the-data`)
+                .set('Authorization', 'Bearer abcd');
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal('http://potato.com/v1/area?page[number]=1&page[size]=1000');
+            response.body.links.should.have.property('prev').and.equal('http://potato.com/v1/area?page[number]=1&page[size]=1000');
+            response.body.links.should.have.property('next').and.equal('http://potato.com/v1/area?page[number]=1&page[size]=1000');
+            response.body.links.should.have.property('first').and.equal('http://potato.com/v1/area?page[number]=1&page[size]=1000');
+            response.body.links.should.have.property('last').and.equal('http://potato.com/v1/area?page[number]=1&page[size]=1000');
+        });
     });
 
     it('Get a page with 3 areas using pagination', async () => {