+++ app/test/e2e/v2/get-areas.spec.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-unused-expressions */
 const nock = require('nock');
 const chai = require('chai');
+const config = require('config');
 const mongoose = require('mongoose');
 
 const Area = require('models/area.modelV2');
@@ -25,6 +26,41 @@ describe('V2 - Get areas', () => {
         if (process.env.NODE_ENV !== 'test') {
             throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);
         }
+    });
+
+    describe('Test pagination links', () => {
+        it('Get areas without referer header should be successful and use the request host', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v2/area`)
+                .set('Authorization', 'Bearer abcd');
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal(`http://127.0.0.1:${config.get('service.port')}/v2/area?page[number]=1&page[size]=300`);
+            response.body.links.should.have.property('prev').and.equal(`http://127.0.0.1:${config.get('service.port')}/v2/area?page[number]=1&page[size]=300`);
+            response.body.links.should.have.property('next').and.equal(`http://127.0.0.1:${config.get('service.port')}/v2/area?page[number]=1&page[size]=300`);
+            response.body.links.should.have.property('first').and.equal(`http://127.0.0.1:${config.get('service.port')}/v2/area?page[number]=1&page[size]=300`);
+            response.body.links.should.have.property('last').and.equal(`http://127.0.0.1:${config.get('service.port')}/v2/area?page[number]=1&page[size]=300`);
+        });
+
+        it('Get areas with referer header should be successful and use that header on the links on the response', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v2/area`)
+                .set('referer', `https://potato.com/get-me-all-the-data`)
+                .set('Authorization', 'Bearer abcd');
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal('http://potato.com/v2/area?page[number]=1&page[size]=300');
+            response.body.links.should.have.property('prev').and.equal('http://potato.com/v2/area?page[number]=1&page[size]=300');
+            response.body.links.should.have.property('next').and.equal('http://potato.com/v2/area?page[number]=1&page[size]=300');
+            response.body.links.should.have.property('first').and.equal('http://potato.com/v2/area?page[number]=1&page[size]=300');
+            response.body.links.should.have.property('last').and.equal('http://potato.com/v2/area?page[number]=1&page[size]=300');
+        });
     });
 
     it('Getting areas without being logged in should return a 401 - "Not logged" error', async () => {